ai-service/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0
numpy==1.24.3

# Speech Recognition
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Speaker Diarization
pyannote.audio==3.1.1
pyannote.core==5.0.0

# Voice Activity Detection
silero-vad==4.0.1

# AI/LLM
openai==1.10.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2

# HTTP Client
httpx==0.26.0
